feat: add support for uploading PPT(x) files
Closes: https://github.com/2anki/server/issues/1482
