--- src/lib/parser/pdf/convertPDFToImages.ts
@@ -1,135 +1,0 @@
-import { writeFile } from 'fs/promises';
-import path from 'path';
-import { execFile } from 'child_process';
-import Workspace from '../WorkSpace';
-import { S3 } from 'aws-sdk';
-
-function getPageCount(pdfPath: string): Promise<number> {
-  return new Promise((resolve, reject) => {
-    const pdfinfoBin =
-      process.platform === 'darwin'
-        ? '/usr/local/bin/pdfinfo'
-        : '/usr/bin/pdfinfo';
-    execFile(pdfinfoBin, [pdfPath], (error, stdout) => {
-      if (error) {
-        reject(new Error('Failed to execute pdfinfo'));
-        return;
-      }
-
-      const pageCount = parseInt(
-        stdout
-          .split('\n')
-          .find((line) => line.startsWith('Pages:'))
-          ?.split(/\s+/)[1] || '0'
-      );
-
-      if (!pageCount) {
-        reject(new Error('Failed to get page count'));
-        return;
-      }
-
-      resolve(pageCount);
-    });
-  });
-}
-
-function convertPage(
-  pdfPath: string,
-  pageNumber: number,
-  totalPageCount: number
-): Promise<string> {
-  const outputFileNameBase = `${pdfPath}-page${pageNumber}`;
-
-  const determinePaddingLength = (pageCount: number): number => {
-    if (pageCount >= 1000) return 4;
-    if (pageCount >= 100) return 3;
-    if (pageCount >= 10) return 2;
-    return 1;
-  };
-
-  const paddedPageNumber = String(pageNumber).padStart(
-    determinePaddingLength(totalPageCount),
-    '0'
-  );
-
-  return new Promise((resolve, reject) => {
-    execFile(
-      'pdftoppm',
-      [
-        '-png',
-        '-f',
-        pageNumber.toString(),
-        '-l',
-        pageNumber.toString(),
-        pdfPath,
-        outputFileNameBase,
-      ],
-      (error) => {
-        if (error) {
-          return reject(
-            new Error(`Failed to convert page ${pageNumber} to PNG`)
-          );
-        }
-        resolve(`${outputFileNameBase}-${paddedPageNumber}.png`);
-      }
-    );
-  });
-}
-
-function combineIntoHTML(imagePaths: string[], title: string): string {
-  const html = `<!DOCTYPE html>
-<html>
-<head><title>${title}</title></head>
-<body>
-  ${Array.from({ length: imagePaths.length / 2 }, (_, i) => {
-    const front = path.basename(imagePaths[i * 2]);
-    const back = path.basename(imagePaths[i * 2 + 1]);
-    return `<ul class="toggle">
-    <li>
-      <details>
-        <summary>
-        <img src="${front}" />
-        </summary>
-        <img src="${back}" />
-      </details>
-    </li>
-    </ul>`;
-  }).join('\n')}
-</body>
-</html>`;
-
-  return html;
-}
-
-interface ConvertPDFToImagesInput {
-  workspace: Workspace;
-  noLimits: boolean;
-  contents?: S3.Body;
-  name?: string;
-}
-
-export const PDF_EXCEEDS_MAX_PAGE_LIMIT =
-  'PDF exceeds maximum page limit of 100 for free and anonymous users.';
-
-export async function convertPDFToImages(
-  input: ConvertPDFToImagesInput
-): Promise<Buffer> {
-  const { contents, workspace, noLimits, name } = input;
-  const pdfPath = path.join(workspace.location, name ?? 'Default.pdf');
-  await writeFile(pdfPath, Buffer.from(contents as Buffer));
-
-  const pageCount = await getPageCount(pdfPath);
-  const title = path.basename(pdfPath);
-  if (!noLimits && pageCount > 100) {
-    throw new Error(PDF_EXCEEDS_MAX_PAGE_LIMIT);
-  }
-
-  const imagePaths = await Promise.all(
-    Array.from({ length: pageCount }, (_, i) =>
-      convertPage(pdfPath, i + 1, pageCount)
-    )
-  );
-
-  const html = combineIntoHTML(imagePaths, title);
-  return Buffer.from(html);
-}
